Title: Update README with CI/CD and Deployment Details
Description:
This commit updates the README.md file with detailed instructions on how to use the CI/CD pipeline for deploying the application. It also includes information on how the application is deployed on AWS using Terraform. The Deployment section has been updated with the steps to deploy changes to the application and the infrastructure.